# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Csíkszereda Programajánló API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  color: string;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  startDate: string;
  endDate?: string | null;
  location: string;
  locationAddress?: string | null;
  categoryId?: number | null;
  category?: Category | null;
  featured: boolean;
  monthHighlight: boolean;
  ticketUrl?: string | null;
  price?: string | null;
  tags: string[];
  /** JSON-encoded news link objects [{title, url}] */
  newsLinks?: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateEventBody {
  title: string;
  description: string;
  imageUrl: string;
  startDate: string;
  endDate?: string | null;
  location: string;
  locationAddress?: string | null;
  categoryId?: number | null;
  featured?: boolean;
  monthHighlight?: boolean;
  ticketUrl?: string | null;
  price?: string | null;
  tags?: string[];
}

export interface UploadUrlRequest {
  /**
   * Original file name.
   * @minLength 1
   */
  name: string;
  /**
   * File size in bytes.
   * @minimum 1
   */
  size: number;
  /**
   * MIME type of the file (e.g. `image/jpeg`).
   * @minLength 1
   */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface Banner {
  id: number;
  title: string;
  imageUrl: string;
  linkUrl?: string | null;
  /** 'full' = teljes szélességű sáv, 'card' = kártyamérű */
  displayType: string;
  active: boolean;
  /** Hány esemény után jelenjen meg */
  position: number;
  createdAt: string;
}

export interface CreateBannerBody {
  title: string;
  imageUrl: string;
  linkUrl?: string | null;
  displayType?: string;
  active?: boolean;
  position?: number;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListEventsParams = {
  categoryId?: number;
  startDate?: string;
  endDate?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type ListEvents200 = {
  events: Event[];
  total: number;
};

export type ListFeaturedEvents200 = {
  events: Event[];
};

export type ListThisWeekEvents200DaysItem = {
  date: string;
  dayName: string;
  events: Event[];
};

export type ListThisWeekEvents200 = {
  days: ListThisWeekEvents200DaysItem[];
};

export type ListUpcomingEventsParams = {
  limit?: number;
};

export type ListUpcomingEvents200 = {
  events: Event[];
};

export type ListBanners200 = {
  banners: Banner[];
};

export type ListAdminBanners200 = {
  banners: Banner[];
};

export type ListCategories200 = {
  categories: Category[];
};