# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Csíkszereda Programajánló API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      description: Returns paginated list of events with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  total:
                    type: integer
                required:
                  - events
                  - total

    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /events/upcoming:
    get:
      operationId: listUpcomingEvents
      tags: [events]
      summary: Get upcoming events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 9
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                required:
                  - events
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
...
  /admin/banners:
    get:
      operationId: listAdminBanners
      tags: [banners]
      summary: List all banners (admin)
      responses:
        "200":
          description: All banners
          content:
            application/json:
              schema:
                type: object
                properties:
                  banners:
                    type: array
                    items:
                      $ref: "#/components/schemas/Banner"
                required:
                  - banners
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "201":
          description: Created banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /admin/banners/{id}:
    patch:
      operationId: updateBanner
      tags: [banners]
      summary: Update banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "200":
          description: Updated banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - slug
        - color
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        location:
          type: string
        locationAddress:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        featured:
          type: boolean
        monthHighlight:
          type: boolean
        ticketUrl:
          type: string
          nullable: true
        price:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        newsLinks:
          type: array
          items:
            type: string
          description: JSON-encoded news link objects [{title, url}]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - imageUrl
        - startDate
        - location
        - featured
        - monthHighlight
        - tags
        - createdAt
        - updatedAt
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        location:
          type: string
        locationAddress:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        featured:
          type: boolean
          default: false
        monthHighlight:
          type: boolean
          default: false
        ticketUrl:
          type: string
          nullable: true
        price:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - imageUrl
        - startDate
        - location
...
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
          nullable: true
        displayType:
          type: string
          description: "'full' = teljes szélességű sáv, 'card' = kártyamérű"
        active:
          type: boolean
        position:
          type: integer
          description: Hány esemény után jelenjen meg
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - imageUrl
        - displayType
        - active
        - position
        - createdAt
    CreateBannerBody:
      type: object
      properties:
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
          nullable: true
        displayType:
          type: string
          default: full
        active:
          type: boolean
          default: true
        position:
          type: integer
          default: 6
      required:
        - title
        - imageUrl
...