# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Csíkszereda Programajánló API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface Event {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  startDate: Date;
  endDate?: Date | null;
  location: string;
  locationAddress?: string | null;
  categoryId?: number | null;
  category?: Category | null;
  featured: boolean;
  monthHighlight: boolean;
  ticketUrl?: string | null;
  price?: string | null;
  tags: string[];
  /** JSON-encoded news link objects [{title, url}] */
  newsLinks?: string[];
  createdAt: Date;
  updatedAt: Date;
}